Moved to submodule, don't trigger if map_unwrap_or does
